fix: silence logic/type and RWX-segment build warnings (cat 3 & 4)
Category 3 (logic / type correctness):
- smp.c: parenthesize TSS selector so (8+2*cpu_id)<<3 computes correctly
  (was 8 + 16*cpu_id due to << binding looser than +).
- trap.c: lseek default sets new_offset=-1 so the common <0 check returns
  -EINVAL instead of an uninitialized read overwriting the error.
- task.c: insert empty stmt after 'retry:' label (c23-ext).
- abort.c: __builtin_unreachable() after kpanic so noreturn holds.
- kallsyms.c: %499s -> %99s (buffer is 100) to fix fortify overflow.
- systest.c: cast time_t/clock_t to long before >=0 (unsigned type-limits).
- pmm.c: zero-init PMMngr with {0}.
- strtol.c: use -(LONG_MAX)-1 instead of LONG_MIN literal.

Category 4 (RWX LOAD segments):
- user/linker.ld: data phdr FLAGS(7)->FLAGS(6) (drop execute).
- kernel/linker.ld: add PHDRS; text/ltext/rodata RX, data/bss RW.

+++ user/systest.c
@@ -443,7 +443,7 @@ static void test_gettimeofday(void)
     struct timeval tv = {0};
     int ret = gettimeofday(&tv, NULL);
     CHECK3(ret == 0, "gettimeofday", "ret 0");
-    CHECKF(tv.tv_sec >= 0, "gettimeofday", "sec=%lu", "sec=%lu", (unsigned long)tv.tv_sec);
+    CHECKF((long)tv.tv_sec >= 0, "gettimeofday", "sec=%lu", "sec=%lu", (unsigned long)tv.tv_sec);
     CHECK3(tv.tv_usec < 1000000ULL, "gettimeofday", "tv_usec < 1e6");
 }
 
@@ -546,7 +546,7 @@ static void test_times(void)
 {
     struct tms buf;
     uint64_t t = times(&buf);
-    CHECK3(t >= 0, "times", "called");
+    CHECK3((long)t >= 0, "times", "called");
 }
 
 // ── 35: uname ──────────────────────────────────────────────
